pyproject.toml
[tool.poetry]
name = "ddbb"
version = "0.1.0"
description = ""
authors = ["user <020200295a@uandina.edu.pe>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
djangorestframework-simplejwt = "^5.2.2"
dj-rest-auth = {version = "4.0.0", extras = ["with-social"]}
django-allauth = "0.52.0"
psycopg2-binary = "^2.9.6"
mssql-django = "^1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
